third_party/WebKit/LayoutTests/fast/dom/document-set-title-no-head.html

<!DOCTYPE html>
<head></head>
<body>
<script src="../../resources/js-test.js"></script>
<script type="text/javascript">
description("Test that setting document.title when there is no &lt;head&gt; or &lt;title&gt; does nothing");
onload = function()
{
    var head = document.head;
    while (head.firstChild)
        document.documentElement.appendChild(head.firstChild);
    document.documentElement.removeChild(head);
    document.title = "New title";
    shouldBeEmptyString('document.title');
}
</script>
</body>
