third_party/WebKit/LayoutTests/fast/dom/document-set-title-no-head-expected.txt

Test that setting document.title when there is no <head> or <title> does nothing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS document.title is ""
